Move ArrayUtils to StoreApi
